Add p4-graphs CLI option -deps-skip-transitive-reduction
Skipping the transitive_reduction calculation can speed things up
enormously for some larger p4 programs we have been working with, with
the disadvantage that if you run GraphViz's dot on the .dot file, the
number of dependencies can then be so large that dot takes too long.
I will add a separate CLI option to skip that step.