Add code to calculate critical paths 'backward' as well as 'forward'
It did 'forward' before, i.e. go through dependency edges in the
forward direction, always trying to find the earliest time that any
activity could be scheduled.

Not it also can do 'backward', i.e. go through dependency edges in the
reverse direction, always trying to find the latest time that any
activity can be scheduled, still satisfying the dependencies.

For activities on the critical path, their earliest and latest
possible scheduled times will be the same.

For activities not on the critical path, those times will be
different, and will show the range that they can be scheduled in
without affecting the critical path.